+++ .eslintrc.cjs
@@ -15,10 +15,10 @@ module.exports = {
   extends: '@adobe/helix',
   overrides: [
     {
-      files: ["*.test.js"],
+      files: ['*.test.js'],
       rules: {
-        "no-unused-expressions": "off"
-      }
-    }
+        'no-unused-expressions': 'off',
+      },
+    },
   ],
 };