+++ test/utils/slack/base.test.js
@@ -15,7 +15,12 @@
 import { expect } from 'chai';
 import sinon from 'sinon';
 
-import { extractURLFromSlackInput, postErrorMessage, sendMessageBlocks } from '../../../src/utils/slack/base.js';
+import {
+  extractURLFromSlackInput,
+  FALLBACK_SLACK_CHANNEL, getSlackChannelId,
+  postErrorMessage,
+  sendMessageBlocks,
+} from '../../../src/utils/slack/base.js';
 
 describe('Base Slack Utils', () => {
   describe('extractBaseURLFromInput', () => {
@@ -25,10 +30,9 @@ describe('Base Slack Utils', () => {
 
     it('extractURLFromSlackInput without path', async () => {
       const expected = 'adobe.com';
-      const expectedWithScheme = `https://${expected}`;
 
       expect(extractURLFromSlackInput('get site adobe.com', false, false)).to.equal(expected);
-      expect(extractURLFromSlackInput('get site <adobe.com|www.adobe.com>', false)).to.equal(expectedWithScheme);
+      expect(extractURLFromSlackInput('get site <adobe.com|www.adobe.com>', false, false)).to.equal(expected);
       expect(extractURLFromSlackInput('get site adobe.com/', false, false)).to.equal(expected);
       expect(extractURLFromSlackInput('get site <adobe.com/>', false, false)).to.equal(expected);
       expect(extractURLFromSlackInput('get site http://adobe.com', false, false)).to.equal(expected);
@@ -44,69 +48,70 @@ describe('Base Slack Utils', () => {
     it('extractURLFromSlackInput with path', async () => {
       const expected = 'adobe.com/some/path/w1th_numb3rs';
 
-      expect(extractURLFromSlackInput('add site http://adobe.com/some/path/w1th_numb3rs', false), expected);
-      expect(extractURLFromSlackInput('add site <http://adobe.com/some/path/w1th_numb3rs|adobe.com/some/path/w1th_numb3rs>', false), expected);
-      expect(extractURLFromSlackInput('add site https://adobe.com/some/path/w1th_numb3rs', false), expected);
-      expect(extractURLFromSlackInput('add site <https://adobe.com/some/path/w1th_numb3rs|adobe.com/some/path/w1th_numb3rs>', false), expected);
-      expect(extractURLFromSlackInput('add site https://www.adobe.com/some/path/w1th_numb3rs', false), expected);
-      expect(extractURLFromSlackInput('add site <https://www.adobe.com/some/path/w1th_numb3rs|www.adobe.com/some/path/w1th_numb3rs>', false), expected);
-      expect(extractURLFromSlackInput('add site https://www.adobe.com/some/path/w1th_numb3rs/', false), `${expected}/`);
-      expect(extractURLFromSlackInput('add site <https://www.adobe.com/some/path/w1th_numb3rs/>', false), `${expected}/`);
+      expect(extractURLFromSlackInput('add site http://adobe.com/some/path/w1th_numb3rs', false, false)).to.equal(expected);
+      expect(extractURLFromSlackInput('add site <http://adobe.com/some/path/w1th_numb3rs|adobe.com/some/path/w1th_numb3rs>', false, false)).to.equal(expected);
+      expect(extractURLFromSlackInput('add site https://adobe.com/some/path/w1th_numb3rs', false, false)).to.equal(expected);
+      expect(extractURLFromSlackInput('add site <https://adobe.com/some/path/w1th_numb3rs|adobe.com/some/path/w1th_numb3rs>', false, false)).to.equal(expected);
+      expect(extractURLFromSlackInput('add site https://www.adobe.com/some/path/w1th_numb3rs', false, false)).to.equal(expected);
+      expect(extractURLFromSlackInput('add site <https://www.adobe.com/some/path/w1th_numb3rs|www.adobe.com/some/path/w1th_numb3rs>', false, false)).to.equal(expected);
+      expect(extractURLFromSlackInput('add site https://www.adobe.com/some/path/w1th_numb3rs/', false, false)).to.equal(`${expected}/`);
+      expect(extractURLFromSlackInput('add site <https://www.adobe.com/some/path/w1th_numb3rs/>', false, false)).to.equal(`${expected}/`);
     });
 
     it('extractURLFromSlackInput with subdomain and path', async () => {
       const expected = 'business.adobe.com/some/path/w1th_numb3rs';
 
-      expect(extractURLFromSlackInput('get site http://business.adobe.com/some/path/w1th_numb3rs', false), expected);
-      expect(extractURLFromSlackInput('get site <http://business.adobe.com/some/path/w1th_numb3rs|business.adobe.com/some/path/w1th_numb3rs>', false), expected);
-      expect(extractURLFromSlackInput('get site https://business.adobe.com/some/path/w1th_numb3rs', false), expected);
-      expect(extractURLFromSlackInput('get site <https://business.adobe.com/some/path/w1th_numb3rs|business.adobe.com/some/path/w1th_numb3rs>', false), expected);
-      expect(extractURLFromSlackInput('add site https://business.adobe.com/some/path/w1th_numb3rs/', false), `${expected}/`);
-      expect(extractURLFromSlackInput('add site <https://business.adobe.com/some/path/w1th_numb3rs/>', false), `${expected}/`);
+      expect(extractURLFromSlackInput('get site http://business.adobe.com/some/path/w1th_numb3rs', false, false)).to.equal(expected);
+      expect(extractURLFromSlackInput('get site <http://business.adobe.com/some/path/w1th_numb3rs|business.adobe.com/some/path/w1th_numb3rs>', false, false)).to.equal(expected);
+      expect(extractURLFromSlackInput('get site https://business.adobe.com/some/path/w1th_numb3rs', false, false)).to.equal(expected);
+      expect(extractURLFromSlackInput('get site <https://business.adobe.com/some/path/w1th_numb3rs|business.adobe.com/some/path/w1th_numb3rs>', false, false)).to.equal(expected);
+      expect(extractURLFromSlackInput('add site https://business.adobe.com/some/path/w1th_numb3rs/', false, false)).to.equal(`${expected}/`);
+      expect(extractURLFromSlackInput('add site <https://business.adobe.com/some/path/w1th_numb3rs/>', false, false)).to.equal(`${expected}/`);
     });
 
     it('extractURLFromSlackInput with subdomain, path and extension', async () => {
-      const expected = 'personal.nedbank.co.za/borrow/personal-loans.html';
+      const expected = 'https://personal.nedbank.co.za/borrow/personal-loans.html';
 
-      expect(extractURLFromSlackInput('get site personal.nedbank.co.za/borrow/personal-loans.html', false), expected);
-      expect(extractURLFromSlackInput('get site <personal.nedbank.co.za/borrow/personal-loans.html|personal.nedbank.co.za/borrow/personal-loans.html>', false), expected);
-      expect(extractURLFromSlackInput('get site https://personal.nedbank.co.za/borrow/personal-loans.html', false), expected);
-      expect(extractURLFromSlackInput('get site <https://personal.nedbank.co.za/borrow/personal-loans.html|personal.nedbank.co.za/borrow/personal-loans.html>', false), expected);
-      expect(extractURLFromSlackInput('get site https://personal.nedbank.co.za/borrow/personal-loans.html/', false), expected);
-      expect(extractURLFromSlackInput('get site <https://personal.nedbank.co.za/borrow/personal-loans.html/>', false), expected);
+      expect(extractURLFromSlackInput('get site personal.nedbank.co.za/borrow/personal-loans.html', false)).to.equal(expected);
+      expect(extractURLFromSlackInput('get site <personal.nedbank.co.za/borrow/personal-loans.html|personal.nedbank.co.za/borrow/personal-loans.html>', false)).to.equal(expected);
+      expect(extractURLFromSlackInput('get site https://personal.nedbank.co.za/borrow/personal-loans.html', false)).to.equal(expected);
+      expect(extractURLFromSlackInput('get site <https://personal.nedbank.co.za/borrow/personal-loans.html|personal.nedbank.co.za/borrow/personal-loans.html>', false)).to.equal(expected);
+      expect(extractURLFromSlackInput('get site https://personal.nedbank.co.za/borrow/personal-loans.html/', false)).to.equal(`${expected}/`);
+      expect(extractURLFromSlackInput('get site <https://personal.nedbank.co.za/borrow/personal-loans.html/>', false)).to.equal(`${expected}/`);
     });
 
     it('extractURLFromSlackInput with subdomain, path, selector and extension', async () => {
-      const expected = 'personal.nedbank.co.za/borrow/personal-loans.plain.html';
+      const expected = 'https://personal.nedbank.co.za/borrow/personal-loans.plain.html';
 
-      expect(extractURLFromSlackInput('get site personal.nedbank.co.za/borrow/personal-loans.plain.html', false), expected);
-      expect(extractURLFromSlackInput('get site <personal.nedbank.co.za/borrow/personal-loans.plain.html|personal.nedbank.co.za/borrow/personal-loans.plain.html>', false), expected);
-      expect(extractURLFromSlackInput('get site https://personal.nedbank.co.za/borrow/personal-loans.plain.html', false), expected);
-      expect(extractURLFromSlackInput('get site <https://personal.nedbank.co.za/borrow/personal-loans.plain.html|personal.nedbank.co.za/borrow/personal-loans.plain.html>', false), expected);
-      expect(extractURLFromSlackInput('get site https://personal.nedbank.co.za/borrow/personal-loans.plain.html/', false), expected);
-      expect(extractURLFromSlackInput('get site <https://personal.nedbank.co.za/borrow/personal-loans.plain.html/>', false), expected);
+      expect(extractURLFromSlackInput('get site personal.nedbank.co.za/borrow/personal-loans.plain.html', false)).to.equal(expected);
+      expect(extractURLFromSlackInput('get site <personal.nedbank.co.za/borrow/personal-loans.plain.html|personal.nedbank.co.za/borrow/personal-loans.plain.html>', false)).to.equal(expected);
+      expect(extractURLFromSlackInput('get site https://personal.nedbank.co.za/borrow/personal-loans.plain.html', false)).to.equal(expected);
+      expect(extractURLFromSlackInput('get site <https://personal.nedbank.co.za/borrow/personal-loans.plain.html|personal.nedbank.co.za/borrow/personal-loans.plain.html>', false)).to.equal(expected);
+      expect(extractURLFromSlackInput('get site https://personal.nedbank.co.za/borrow/personal-loans.plain.html/', false)).to.equal(`${expected}/`);
+      expect(extractURLFromSlackInput('get site <https://personal.nedbank.co.za/borrow/personal-loans.plain.html/>', false)).to.equal(`${expected}/`);
     });
 
     it('extractURLFromSlackInput domain only', async () => {
       const expected = 'business.adobe.com';
 
-      expect(extractURLFromSlackInput('get site http://business.adobe.com/some/path/w1th_numb3rs', true), expected);
-      expect(extractURLFromSlackInput('get site <http://business.adobe.com/some/path/w1th_numb3rs|business.adobe.com/some/path/w1th_numb3rs>', true), expected);
-      expect(extractURLFromSlackInput('get site https://business.adobe.com/some/path/w1th_numb3rs', true), expected);
-      expect(extractURLFromSlackInput('get site <https://business.adobe.com/some/path/w1th_numb3rs|business.adobe.com/some/path/w1th_numb3rs>', true), expected);
-      expect(extractURLFromSlackInput('add site https://business.adobe.com/some/path/w1th_numb3rs/', true), expected);
-      expect(extractURLFromSlackInput('add site <https://business.adobe.com/some/path/w1th_numb3rs/>', true), expected);
+      expect(extractURLFromSlackInput('get site business.adobe.com/', true)).to.equal(expected);
+      expect(extractURLFromSlackInput('get site http://business.adobe.com/some/path/w1th_numb3rs', true)).to.equal(expected);
+      expect(extractURLFromSlackInput('get site <http://business.adobe.com/some/path/w1th_numb3rs|business.adobe.com/some/path/w1th_numb3rs>', true)).to.equal(expected);
+      expect(extractURLFromSlackInput('get site https://business.adobe.com/some/path/w1th_numb3rs', true)).to.equal(expected);
+      expect(extractURLFromSlackInput('get site <https://business.adobe.com/some/path/w1th_numb3rs|business.adobe.com/some/path/w1th_numb3rs>', true)).to.equal(expected);
+      expect(extractURLFromSlackInput('add site https://business.adobe.com/some/path/w1th_numb3rs/', true)).to.equal(expected);
+      expect(extractURLFromSlackInput('add site <https://business.adobe.com/some/path/w1th_numb3rs/>', true)).to.equal(expected);
     });
 
     it('extractURLFromSlackInput with trailing tokens', async () => {
-      const expected = 'personal.nedbank.co.za/borrow/personal-loans.plain.html';
+      const expected = 'https://personal.nedbank.co.za/borrow/personal-loans.plain.html';
 
-      expect(extractURLFromSlackInput('get site personal.nedbank.co.za/borrow/personal-loans.plain.html test', false), expected);
-      expect(extractURLFromSlackInput('get site <personal.nedbank.co.za/borrow/personal-loans.plain.html|personal.nedbank.co.za/borrow/personal-loans.plain.html> test', false), expected);
-      expect(extractURLFromSlackInput('get site https://personal.nedbank.co.za/borrow/personal-loans.plain.html www.acme.com', false), expected);
-      expect(extractURLFromSlackInput('get site <https://personal.nedbank.co.za/borrow/personal-loans.plain.html|personal.nedbank.co.za/borrow/personal-loans.plain.html> www.acme.com', false), expected);
-      expect(extractURLFromSlackInput('get site https://personal.nedbank.co.za/borrow/personal-loans.plain.html/ extra acme.com/', false), expected);
-      expect(extractURLFromSlackInput('get site <https://personal.nedbank.co.za/borrow/personal-loans.plain.html/> extra acme.com/ <acme.com/> <http://acme.com|acme.com>', false), expected);
+      expect(extractURLFromSlackInput('get site personal.nedbank.co.za/borrow/personal-loans.plain.html test', false)).to.equal(expected);
+      expect(extractURLFromSlackInput('get site <personal.nedbank.co.za/borrow/personal-loans.plain.html|personal.nedbank.co.za/borrow/personal-loans.plain.html> test', false)).to.equal(expected);
+      expect(extractURLFromSlackInput('get site https://personal.nedbank.co.za/borrow/personal-loans.plain.html www.acme.com', false)).to.equal(expected);
+      expect(extractURLFromSlackInput('get site <https://personal.nedbank.co.za/borrow/personal-loans.plain.html|personal.nedbank.co.za/borrow/personal-loans.plain.html> www.acme.com', false)).to.equal(expected);
+      expect(extractURLFromSlackInput('get site https://personal.nedbank.co.za/borrow/personal-loans.plain.html/ extra acme.com/', false)).to.equal(`${expected}/`);
+      expect(extractURLFromSlackInput('get site <https://personal.nedbank.co.za/borrow/personal-loans.plain.html/> extra acme.com/ <acme.com/> <http://acme.com|acme.com>', false)).to.equal(`${expected}/`);
     });
   });
 
@@ -160,6 +165,28 @@ describe('Base Slack Utils', () => {
 
         expect(say.calledOnce).to.be.true;
         expect(say.firstCall.args[0]).to.deep.equal({ blocks: expectedBlocks });
+      });
+    });
+    describe('getSlackChannelId()', () => {
+      it('fallbacks to default slack channel when no configured', async () => {
+        expect(getSlackChannelId(undefined, undefined)).to.equal(FALLBACK_SLACK_CHANNEL);
+        expect(getSlackChannelId(null, '')).to.equal(FALLBACK_SLACK_CHANNEL);
+        expect(getSlackChannelId(undefined, '')).to.equal(FALLBACK_SLACK_CHANNEL);
+        expect(getSlackChannelId(undefined, ',')).to.equal(FALLBACK_SLACK_CHANNEL);
+        expect(getSlackChannelId(undefined, '=,')).to.equal(FALLBACK_SLACK_CHANNEL);
+        expect(getSlackChannelId(undefined, 'ch= ,')).to.equal(FALLBACK_SLACK_CHANNEL);
+        expect(getSlackChannelId('channel', 'channel1=,channel2= ')).to.equal(FALLBACK_SLACK_CHANNEL);
+        expect(getSlackChannelId(null, 'asd')).to.equal(FALLBACK_SLACK_CHANNEL);
+      });
+
+      it('fallbacks to default slack channel when no found in env', async () => {
+        const channelId = getSlackChannelId('ch4', 'ch1=ASD,ch2=DSA,ch3=TRE');
+        expect(channelId).to.equal(FALLBACK_SLACK_CHANNEL);
+      });
+
+      it('returs to desired slack channel', async () => {
+        const channelId = getSlackChannelId('ch3', 'ch1=ASD,ch2=DSA,ch3=TRE');
+        expect(channelId).to.equal('TRE');
       });
     });
   });